requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
scikit-learn==1.3.0
openpyxl==3.1.2
PyYAML==6.0.1
reportlab==4.0.4